tests/wpt/tests/css/css-tables/parsing/caption-side-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Table Module: getComputedStyle().captionSide</title>
<link rel="help" href="https://drafts.csswg.org/css-tables/#propdef-caption-side">
<meta name="assert" content="caption-side computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("caption-side", "top");
test_computed_value("caption-side", "bottom");
</script>
</body>
</html>
